add list only flag, displays commands only

+++ docker-course.sh
@@ -96,8 +96,10 @@ echo --------------------------------1.01: Hello world--------------------------
 
 echo -n "$ sudo docker run ubuntu:14.04 /bin/echo 'Hello world'"
 wait_for_keypress;
-
-docker run ubuntu:14.04 /bin/echo 'Hello world'
+if [[ ! $LIST == '1' ]];
+then
+    docker run ubuntu:14.04 /bin/echo 'Hello world'
+fi
 
 # An interactive container
 # -i, --interactive=false    Keep STDIN open even if not attached
@@ -126,19 +128,31 @@ docker run --name=insane_babbage -d ubuntu:14.04 /bin/sh -c "while true; do echo
 
 echo -n '$ sudo docker ps'
 wait_for_keypress;
-docker ps
+if [[ ! $LIST == '1' ]];
+then
+    docker ps
+fi
 
 echo -n '$ sudo docker logs insane_babbage'
 wait_for_keypress;
-docker logs insane_babbage | tail -n 10
+if [[ ! $LIST == '1' ]];
+then
+    docker logs insane_babbage | tail -n 10
+fi
 
 echo -n '$ sudo docker stop insane_babbage'
 wait_for_keypress;
-docker stop insane_babbage
+if [[ ! $LIST == '1' ]];
+then
+    docker stop insane_babbage
+fi
 
 echo -n '$ sudo docker ps'
 wait_for_keypress;
-docker ps
+if [[ ! $LIST == '1' ]];
+then
+    docker ps
+fi
 
 if [[ $BREAK == '1' ]];
 then
@@ -157,7 +171,10 @@ echo --------------------------------2.01: Working with containers--------------
 
 echo -n "$ sudo docker version"
 wait_for_keypress;
-docker version
+if [[ ! $LIST == '1' ]];
+then
+    docker version
+fi
 
 # Get Docker command help
 echo
@@ -165,20 +182,29 @@ echo --------------------------------2.02: Get Docker command help--------------
 
 echo -n "$ docker --help"
 wait_for_keypress;
-docker --help
+if [[ ! $LIST == '1' ]];
+then
+    docker --help
+fi
 echo -n "$ sudo docker attach --help"
 wait_for_keypress;
-docker attach --help
+if [[ ! $LIST == '1' ]];
+then
+    docker attach --help
+fi
 
 # Running a web application in Docker
 echo
 echo --------------------------------2.03: Running a web application in Docker---------------------------------
 
-# Note the use of --name=nostalgic_morse which gives us a consistent name
-# -P, --publish-all=false    Publish all exposed ports to random ports
+echo "# Note the use of --name=nostalgic_morse which gives us a consistent name"
+echo "# -P, --publish-all=false    Publish all exposed ports to random ports"
 echo -n '$ sudo docker run --name=nostalgic_morse -d -P training/webapp python app.py'
 wait_for_keypress;
-docker run --name=nostalgic_morse -d -P training/webapp python app.py
+if [[ ! $LIST == '1' ]];
+then
+    docker run --name=nostalgic_morse -d -P training/webapp python app.py
+fi
 
 # Viewing our web application container
 echo
@@ -186,7 +212,10 @@ echo --------------------------------2.04: Viewing our web application container
 
 echo -n '$ sudo docker ps -l'
 wait_for_keypress;
-docker ps -l
+if [[ ! $LIST == '1' ]];
+then
+    docker ps -l
+fi
 
 # A network port shortcut
 echo
@@ -194,18 +223,27 @@ echo --------------------------------2.05: A network port shortcut--------------
 
 echo -n "$ sudo docker port nostalgic_morse 5000"
 wait_for_keypress;
-docker port nostalgic_morse 5000
+if [[ ! $LIST == '1' ]];
+then
+    docker port nostalgic_morse 5000
+fi
 
 # 0.0.0.0, in this context, means "all IPv4 addresses on the local machine"
 # Lets take a closer look:
 echo -n "$ curl -s -i http://\$( sudo docker port nostalgic_morse 5000 )/"
 wait_for_keypress;
-curl -s -i http://$( docker port nostalgic_morse 5000 )/
+if [[ ! $LIST == '1' ]];
+then
+    curl -s -i http://$( docker port nostalgic_morse 5000 )/
+fi
 echo
 
 echo -n "$ sudo lsof -i | egrep 'docker|COMMAND'"
 wait_for_keypress;
-lsof -i | egrep 'docker|COMMAND'
+if [[ ! $LIST == '1' ]];
+then
+    lsof -i | egrep 'docker|COMMAND'
+fi
 
 # Viewing the web application’s logs
 echo
@@ -226,7 +264,10 @@ echo --------------------------------2.07: Looking at our web application contai
 
 echo -n "$ sudo docker top nostalgic_morse"
 wait_for_keypress;
-docker top nostalgic_morse
+if [[ ! $LIST == '1' ]];
+then
+    docker top nostalgic_morse
+fi
 
 # Inspecting our web application container
 echo
@@ -234,11 +275,17 @@ echo --------------------------------2.08: Inspecting our web application contai
 
 echo -n "$ sudo docker inspect nostalgic_morse"
 wait_for_keypress;
-docker inspect nostalgic_morse
+if [[ ! $LIST == '1' ]];
+then
+    docker inspect nostalgic_morse
+fi
 
 echo -n "$ sudo docker inspect -f '{{ .NetworkSettings.IPAddress }}' nostalgic_morse"
 wait_for_keypress;
-docker inspect -f '{{ .NetworkSettings.IPAddress }}' nostalgic_morse
+if [[ ! $LIST == '1' ]];
+then
+    docker inspect -f '{{ .NetworkSettings.IPAddress }}' nostalgic_morse
+fi
 
 # Stopping our web application container
 echo
@@ -246,11 +293,17 @@ echo --------------------------------2.09: Stopping our web application containe
 
 echo -n "$ sudo docker stop nostalgic_morse"
 wait_for_keypress;
-docker stop nostalgic_morse
+if [[ ! $LIST == '1' ]];
+then
+    docker stop nostalgic_morse
+fi
 
 echo -n "$ sudo docker ps -l"
 wait_for_keypress;
-docker ps -l
+if [[ ! $LIST == '1' ]];
+then
+    docker ps -l
+fi
 
 # Restarting our web application container
 echo
@@ -258,11 +311,17 @@ echo --------------------------------2.10: Restarting our web application contai
 
 echo -n "$ sudo docker start nostalgic_morse"
 wait_for_keypress;
-docker start nostalgic_morse
+if [[ ! $LIST == '1' ]];
+then
+    docker start nostalgic_morse
+fi
 
 echo -n "$ sudo docker ps -l"
 wait_for_keypress;
-docker ps -l
+if [[ ! $LIST == '1' ]];
+then
+    docker ps -l
+fi
 
 # also 'sudo docker restart nostalgic_morse'
 
@@ -273,15 +332,24 @@ echo --------------------------------2.11: Removing our web application containe
 echo "# we expect this command to fail as the container is still running"
 echo -n "$ sudo docker rm nostalgic_morse"
 wait_for_keypress;
-docker rm nostalgic_morse
+if [[ ! $LIST == '1' ]];
+then
+    docker rm nostalgic_morse
+fi
 
 echo -n "$ sudo docker stop nostalgic_morse"
 wait_for_keypress;
-docker stop nostalgic_morse
+if [[ ! $LIST == '1' ]];
+then
+    docker stop nostalgic_morse
+fi
 
 echo -n "$ sudo docker rm nostalgic_morse"
 wait_for_keypress;
-docker rm nostalgic_morse
+if [[ ! $LIST == '1' ]];
+then
+    docker rm nostalgic_morse
+fi
 
 if [[ $BREAK == '2' ]];
 then
@@ -300,7 +368,10 @@ echo --------------------------------3.01: Listing images on the host-----------
 
 echo -n '$ sudo docker images'
 wait_for_keypress;
-docker images
+if [[ ! $LIST == '1' ]];
+then
+    docker images
+fi
 
 # Getting a new image
 echo
@@ -308,7 +379,10 @@ echo --------------------------------3.02: Getting a new image------------------
 
 echo -n '$ sudo docker pull centos'
 wait_for_keypress;
-docker pull centos
+if [[ ! $LIST == '1' ]];
+then
+    docker pull centos
+fi
 
 echo "# type exit when you are done looking around"
 echo -n '$ sudo docker run -t -i centos /bin/bash'
@@ -324,7 +398,10 @@ echo --------------------------------3.03: Finding images-----------------------
 
 echo -n '$ sudo docker search sinatra'
 wait_for_keypress;
-docker search sinatra
+if [[ ! $LIST == '1' ]];
+then
+    docker search sinatra
+fi
 
 # Pulling our image
 echo
@@ -332,7 +409,10 @@ echo --------------------------------3.04: Pulling our image--------------------
 
 echo -n '$ sudo docker pull training/sinatra'
 wait_for_keypress;
-docker pull training/sinatra
+if [[ ! $LIST == '1' ]];
+then
+    docker pull training/sinatra
+fi
 
 echo "# type exit when you are done looking around"
 echo -n '$ sudo docker run -t -i training/sinatra /bin/bash'
@@ -354,18 +434,27 @@ if [[ $INTERACTIVE == '1' ]];
 then
     docker run -t -i training/sinatra /bin/bash
 else
-    docker run training/sinatra gem install json
+    if [[ ! $LIST == '1' ]];
+    then
+        docker run training/sinatra gem install json
+    fi
 fi
 
 CONTAINER_ID=$(docker ps -l -q)
 
 echo -n "$ sudo docker commit -m 'Added json gem' -a \"$DOCKER_AUTHOR\" $CONTAINER_ID $OUR_USER/sinatra:v1"
 wait_for_keypress;
-docker commit -m "Added json gem" -a "$DOCKER_AUTHOR" $CONTAINER_ID $OUR_USER/sinatra:v1
+if [[ ! $LIST == '1' ]];
+then
+    docker commit -m "Added json gem" -a "$DOCKER_AUTHOR" $CONTAINER_ID $OUR_USER/sinatra:v1
+fi
 
 echo -n '$ sudo docker images'
 wait_for_keypress;
-docker images
+if [[ ! $LIST == '1' ]];
+then
+    docker images
+fi
 
 echo "# type exit when you are done"
 echo -n "$ sudo docker run -t -i $OUR_USER/sinatra:v1 /bin/bash"
@@ -388,36 +477,51 @@ if [ ! -d "$DIRECTORY" ]; then
 fi
 echo -n "$ cd $DIRECTORY"
 wait_for_keypress;
-cd $DIRECTORY
+if [[ ! $LIST == '1' ]];
+then
+    cd $DIRECTORY
+fi
 echo -n "$ touch $DOCKERFILE"
 wait_for_keypress;
 
-if [ -f "$DOCKERFILE" ]
+if [[ ! $LIST == '1' ]];
 then
-    rm $DOCKERFILE
-else
-    touch $DOCKERFILE
+    if [ -f "$DOCKERFILE" ]
+    then
+        rm $DOCKERFILE
+    else
+        touch $DOCKERFILE
+    fi
 fi
 
 if [[ $INTERACTIVE == '1' ]];
 then
     $EDITOR Dockerfile
 else
-    echo "# This is a comment
+    if [[ ! $LIST == '1' ]];
+    then
+        echo "# This is a comment
 FROM ubuntu:14.04
 MAINTAINER $DOCKER_AUTHOR <$DOCKER_EMAIL>
 RUN apt-get update && apt-get install -y ruby ruby-dev
 RUN gem install sinatra
 " > Dockerfile
+    fi
 fi
 
 echo -n "$ cat Dockerfile"
 wait_for_keypress;
-cat Dockerfile
+if [[ ! $LIST == '1' ]];
+then
+    cat Dockerfile
+fi
 
 echo -n "$ sudo docker build -t $OUR_USER/sinatra:v2 ."
 wait_for_keypress;
-docker build -t $OUR_USER/sinatra:v2 .
+if [[ ! $LIST == '1' ]];
+then
+    docker build -t $OUR_USER/sinatra:v2 .
+fi
 
 # Setting tags on an image
 echo
@@ -426,14 +530,20 @@ echo --------------------------------3.07: Setting tags on an image-------------
 IMAGE_ID=$( docker images $OUR_USER/sinatra  | grep v2 | awk '{print $3}' )
 echo -n "$ sudo docker tag $IMAGE_ID $OUR_USER/sinatra:devel"
 wait_for_keypress;
-if [ ! docker images | awk '{ print $1 $2 }' | grep -q "$OUR_USER/sinatradevel" ]
+if [[ ! $LIST == '1' ]];
 then
-    docker tag $IMAGE_ID $OUR_USER/sinatra:devel
+    if [ ! docker images | awk '{ print $1 $2 }' | grep -q "$OUR_USER/sinatradevel" ]
+    then
+        docker tag $IMAGE_ID $OUR_USER/sinatra:devel
+    fi
 fi
 
 echo -n "$ sudo docker images $OUR_USER/sinatra"
 wait_for_keypress;
-docker images $OUR_USER/sinatra
+if [[ ! $LIST == '1' ]];
+then
+    docker images $OUR_USER/sinatra
+fi
 
 # Image Digests
 echo
@@ -441,7 +551,10 @@ echo --------------------------------3.08: Image Digests------------------------
 
 echo -n "$ sudo docker images --digests | head"
 wait_for_keypress;
-docker images --digests | head
+if [[ ! $LIST == '1' ]];
+then
+    docker images --digests | head
+fi
 
 # Push an image to Docker Hub
 echo
@@ -454,7 +567,10 @@ echo --------------------------------3.10: Remove an image from the host--------
 
 echo -n "$ sudo docker rmi training/sinatra"
 wait_for_keypress;
-docker rmi training/sinatra
+if [[ ! $LIST == '1' ]];
+then
+    docker rmi training/sinatra
+fi
 
 if [[ $BREAK == '3' ]];
 then
@@ -474,11 +590,17 @@ echo --------------------------------4.01: The importance of naming-------------
 
 echo -n "$ sudo docker run -d -P --name web training/webapp python app.py"
 wait_for_keypress;
-docker run -d -P --name web training/webapp python app.py
+if [[ ! $LIST == '1' ]];
+then
+    docker run -d -P --name web training/webapp python app.py
+fi
 
 echo -n "$ sudo docker ps -l"
 wait_for_keypress;
-docker ps -l
+if [[ ! $LIST == '1' ]];
+then
+    docker ps -l
+fi
 
 # Communication across links
 echo
@@ -486,23 +608,38 @@ echo --------------------------------4.02: Communication across links-----------
 
 echo -n "$ sudo docker run -d --name db training/postgres"
 wait_for_keypress;
-docker run -d --name db training/postgres
+if [[ ! $LIST == '1' ]];
+then
+    docker run -d --name db training/postgres
+fi
 
 echo -n "$ sudo docker rm -f web"
 wait_for_keypress;
-docker rm -f web
+if [[ ! $LIST == '1' ]];
+then
+    docker rm -f web
+fi
 
 echo -n "$ sudo docker run -d -P --name web --link db:db training/webapp python app.py"
 wait_for_keypress;
-docker run -d -P --name web --link db:db training/webapp python app.py
+if [[ ! $LIST == '1' ]];
+then
+    docker run -d -P --name web --link db:db training/webapp python app.py
+fi
 
 echo -n "$ sudo docker inspect -f "{{ .HostConfig.Links }}" web"
 wait_for_keypress;
-docker inspect -f "{{ .HostConfig.Links }}" web
+if [[ ! $LIST == '1' ]];
+then
+    docker inspect -f "{{ .HostConfig.Links }}" web
+fi
 
 echo -n "$ sudo docker run --rm --name web2 --link db:db training/webapp env"
 wait_for_keypress;
-docker run --rm --name web2 --link db:db training/webapp env
+if [[ ! $LIST == '1' ]];
+then
+    docker run --rm --name web2 --link db:db training/webapp env
+fi
 
 # Updating the /etc/hosts file
 echo
@@ -519,12 +656,18 @@ if [[ $INTERACTIVE == '1' ]];
 then
     docker run -t -i --rm --link db:webdb training/webapp /bin/bash
 else
-    docker run --rm --link db:webdb training/webapp cat /etc/hosts
+    if [[ ! $LIST == '1' ]];
+    then
+        docker run --rm --link db:webdb training/webapp cat /etc/hosts
+    fi
 fi
 
 echo -n "$ sudo docker restart db"
 wait_for_keypress;
-docker restart db
+if [[ ! $LIST == '1' ]];
+then
+    docker restart db
+fi
 
 echo "# type 'cat /etc/hosts'"
 echo "# type exit when you are done"
@@ -534,7 +677,10 @@ if [[ $INTERACTIVE == '1' ]];
 then
     docker run -t -i --rm --link db:db training/webapp /bin/bash
 else
-    docker run --rm --link db:db training/webapp cat /etc/hosts
+    if [[ ! $LIST == '1' ]];
+    then
+        docker run --rm --link db:db training/webapp cat /etc/hosts
+    fi
 fi
 
 if [[ $BREAK == '4' ]];
@@ -550,11 +696,17 @@ wait_for_keypress;
 echo
 echo --------------------------------5.01: Adding a data volume---------------------------------
 
-docker rename web web1
+if [[ ! $LIST == '1' ]];
+then
+    docker rename web web1
+fi
 
 echo -n "$ sudo docker run -d -P --name web -v /webapp training/webapp python app.py"
 wait_for_keypress;
-docker run -d -P --name web -v /webapp training/webapp python app.py
+if [[ ! $LIST == '1' ]];
+then
+    docker run -d -P --name web -v /webapp training/webapp python app.py
+fi
 
 # Locating a volume
 echo
@@ -562,23 +714,38 @@ echo --------------------------------5.02: Locating a volume--------------------
 
 echo -n "$ sudo docker inspect web"
 wait_for_keypress;
-docker inspect web
+if [[ ! $LIST == '1' ]];
+then
+    docker inspect web
+fi
 
 # Mount a host directory as a data volume
 echo
 echo --------------------------------5.03: Mount a host directory as a data volume---------------------------------
 
-docker rename web web2
+if [[ ! $LIST == '1' ]];
+then
+    docker rename web web2
+fi
 
 echo -n "$ sudo docker run -d -P --name web -v /src/webapp:/opt/webapp training/webapp python app.py"
 wait_for_keypress;
-docker run -d -P --name web -v /src/webapp:/opt/webapp training/webapp python app.py
+if [[ ! $LIST == '1' ]];
+then
+    docker run -d -P --name web -v /src/webapp:/opt/webapp training/webapp python app.py
+fi
 
-docker rename web web3
+if [[ ! $LIST == '1' ]];
+then
+    docker rename web web3
+fi
 
 echo -n "$ sudo docker run -d -P --name web -v /src/webapp:/opt/webapp:ro training/webapp python app.py"
 wait_for_keypress;
-docker run -d -P --name web -v /src/webapp:/opt/webapp:ro training/webapp python app.py
+if [[ ! $LIST == '1' ]];
+then
+    docker run -d -P --name web -v /src/webapp:/opt/webapp:ro training/webapp python app.py
+fi
 
 # Mount a host file as a data volume
 echo
@@ -598,19 +765,31 @@ echo --------------------------------5.05: Creating and mounting a data volume c
 
 echo -n "$ sudo docker create -v /dbdata --name dbdata training/postgres /bin/true"
 wait_for_keypress;
-docker create -v /dbdata --name dbdata training/postgres /bin/true
+if [[ ! $LIST == '1' ]];
+then
+    docker create -v /dbdata --name dbdata training/postgres /bin/true
+fi
 
 echo -n "$ sudo docker run -d --volumes-from dbdata --name db1 training/postgres"
 wait_for_keypress;
-docker run -d --volumes-from dbdata --name db1 training/postgres
+if [[ ! $LIST == '1' ]];
+then
+    docker run -d --volumes-from dbdata --name db1 training/postgres
+fi
 
 echo -n "$ sudo docker run -d --volumes-from dbdata --name db2 training/postgres"
 wait_for_keypress;
-docker run -d --volumes-from dbdata --name db2 training/postgres
+if [[ ! $LIST == '1' ]];
+then
+    docker run -d --volumes-from dbdata --name db2 training/postgres
+fi
 
 echo -n "$ sudo docker run -d --name db3 --volumes-from db1 training/postgres"
 wait_for_keypress;
-docker run -d --name db3 --volumes-from db1 training/postgres
+if [[ ! $LIST == '1' ]];
+then
+    docker run -d --name db3 --volumes-from db1 training/postgres
+fi
 
 # Backup, restore, or migrate data volumes
 echo
@@ -618,16 +797,25 @@ echo --------------------------------5.06: Backup, restore, or migrate data volu
 
 echo -n "$ sudo docker run --volumes-from dbdata -v $(pwd):/backup ubuntu tar cvf /backup/backup.tar /dbdata"
 wait_for_keypress;
-docker run --volumes-from dbdata -v $(pwd):/backup ubuntu tar cvf /backup/backup.tar /dbdata
+if [[ ! $LIST == '1' ]];
+then
+    docker run --volumes-from dbdata -v $(pwd):/backup ubuntu tar cvf /backup/backup.tar /dbdata
+fi
 
 echo -n "$ sudo docker run -v /dbdata --name dbdata2 ubuntu /bin/bash"
 wait_for_keypress;
-docker run -v /dbdata --name dbdata2 ubuntu /bin/bash
-# why no shell here?
+if [[ ! $LIST == '1' ]];
+then
+    docker run -v /dbdata --name dbdata2 ubuntu /bin/bash
+    # why no shell here?
+fi
 
 echo -n "$ sudo docker run --volumes-from dbdata2 -v $(pwd):/backup ubuntu cd /dbdata && tar xvf /backup/backup.tar"
 wait_for_keypress;
-docker run --volumes-from dbdata2 -v $(pwd):/backup ubuntu cd /dbdata && tar xvf /backup/backup.tar
+if [[ ! $LIST == '1' ]];
+then
+    docker run --volumes-from dbdata2 -v $(pwd):/backup ubuntu cd /dbdata && tar xvf /backup/backup.tar
+fi
 
 if [[ $BREAK == '5' ]];
 then
@@ -644,7 +832,10 @@ echo --------------------------------6.01: Searching for images-----------------
 
 echo -n "$ sudo docker search centos"
 wait_for_keypress;
-docker search centos
+if [[ ! $LIST == '1' ]];
+then
+    docker search centos
+fi
 
 if [[ $BREAK == '6' ]];
 then
